Added caching of primer search datastruct
